nodejs: refactor + fix node 21 (#8)
Why
===

after doing the python overlay i realized that i was indeed managing the
node overlay wrong. this new design aligns more with some other overlays
like fenix and zig-overlay.

further, i realized that it's better to keep the original `common.nix`
file from nixpkgs as much as possible and to rely on `callPackage` and
the returned function from `common.nix` to influence builds. this also
means that i can properly support the patches that are in the nixpkgs
repo now

What changed
============

- the `gen.ts` script has been split with the new `nodejs/gen.ts` script
which handles *only* generation of the `versions.json` for the node
overlay
- simplified the structure of the `nodejs` overlay dir
- add patches from nixpkgs when building

Test plan
=========

`nix run '.#nodeVersions."21.7.1"' -- --version` prints the right
version